% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dataframe1}
\alias{dataframe1}
\title{This is the processed dataframe 1 used in the paper (See Table in Annex)}
\format{
An object of class \code{data.frame} with 248 rows and 151 columns.
}
\usage{
dataframe1
}
\description{
This is the processed dataframe 1 used in the paper (See Table in Annex)
}
\author{
World Bank
}
\keyword{Database}
\keyword{Indicators}
\keyword{World}
