% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dataframe2}
\alias{dataframe2}
\title{This is the processed dataframe 2 used in the paper (See Table in Annex)}
\format{
An object of class \code{data.frame} with 216 rows and 213 columns.
}
\usage{
dataframe2
}
\description{
This is the processed dataframe 2 used in the paper (See Table in Annex)
}
\author{
World Bank
}
\keyword{Database}
\keyword{Indicators}
\keyword{World}
