% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dataset1}
\alias{dataset1}
\title{This is the processed data set 1 used in the paper (See Table in Appendix)}
\format{
An object of class \code{data.frame} with 248 rows and 151 columns.
}
\usage{
dataset1
}
\description{
This is the processed data set 1 used in the paper (See Table in Appendix)
}
\author{
World Bank
}
\keyword{Database}
\keyword{Indicators}
\keyword{World}
