% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dataset13}
\alias{dataset13}
\title{This is the processed data set 13 used in the paper (See Table in Appendix)}
\format{
An object of class \code{data.frame} with 85 rows and 559 columns.
}
\usage{
dataset13
}
\description{
This is the processed data set 13 used in the paper (See Table in Appendix)
}
\author{
World Bank
}
\keyword{Database}
\keyword{Indicators}
\keyword{World}
