% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dataset2}
\alias{dataset2}
\title{This is the processed data set 2 used in the paper (See Table in Appendix)}
\format{
An object of class \code{data.frame} with 216 rows and 213 columns.
}
\usage{
dataset2
}
\description{
This is the processed data set 2 used in the paper (See Table in Appendix)
}
\author{
World Bank
}
\keyword{Database}
\keyword{Indicators}
\keyword{World}
