% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalMPI.prediction.using.WID-package.R
\docType{package}
\name{globalMPI.prediction.using.WID-package}
\alias{globalMPI.prediction.using.WID}
\alias{globalMPI.prediction.using.WID-package}
\title{globalMPI.prediction.using.WID: Global MPI Prediction Using World Bank Development Indicators}
\description{
This package includes the code for reproducing the results of the paper.
}
\author{
\strong{Maintainer}: Ignacio Girela \email{ignacio.girela@unc.edi.ar} (\href{https://orcid.org/0000-0003-3297-3854}{ORCID})

}
\keyword{internal}
