% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{globalMPI_all_years_raw}
\alias{globalMPI_all_years_raw}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 444 rows and 7 columns.
}
\usage{
globalMPI_all_years_raw
}
\description{
This is data to be included in my package
}
\references{
\url{data_blah.com}
}
\author{
OPHI-UNDP
}
\keyword{Multidimensional}
\keyword{Poverty}
