% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{Kfold_idxs}
\alias{Kfold_idxs}
\title{Create K-fold indices for Cross Validation}
\usage{
Kfold_idxs(y, k, ...)
}
\arguments{
\item{y}{vector of observations}

\item{k}{number of folds}

\item{...}{}
}
\value{
a vector with length(y) with the corresponding fold each observations belongs
}
\description{
Create K-fold indices for Cross Validation
}
