% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{main_function}
\alias{main_function}
\title{Function used for experiment 1 and 2}
\usage{
main_function(
  data = NULL,
  target = c("MPI", "H", "A"),
  nfolds = 5,
  methods = c("elasticnet", "betaboost"),
  split_size = 0.8,
  ...
)
}
\arguments{
\item{data}{Which of the data frames in Annex is preferred to used}

\item{nfolds}{Number of folds used in K-Fold Cross-Validation for hyperparameter selection}

\item{methods}{Which of the methods to be used ("elasticnet","betaboost")}

\item{split_size}{Proportion of the split of train and test data}

\item{...}{other arguments}

\item{targetTarget}{variable ("MPI","H","A")}
}
\value{
a data frame with the prediction of each method and its ground truth
}
\description{
Function used for experiment 1 and 2
}
