% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_exp1.R
\name{main_function_exp1}
\alias{main_function_exp1}
\title{Function used for experiment 1}
\usage{
main_function_exp1(
  data = NULL,
  target = c("MPI", "H", "A"),
  nfolds = 5,
  methods = c("linear-pls", "beta-pls", "beta-tree-pls", "elasticnet", "beta-elastic",
    "beta-tree-elastic", "betaboost", "xgboost"),
  split_size = 0.8,
  ...
)
}
\arguments{
\item{data}{Which of the data frames in Annex is preferred to used}

\item{target}{Target variable ("MPI","H","A")}

\item{nfolds}{Number of folds used in K-Fold Cross-Validation for hyperparameter selection}

\item{methods}{Which of the methods to be used ("linear-pls","beta-pls","beta-tree-pls","elasticnet","beta-elastic","beta-tree-elastic","betaboost","xgboost")}

\item{split_size}{Proportion of the split of train and test data}

\item{...}{other arguments}
}
\value{
a data frame with the prediction of each method and its ground truth
}
\description{
Function used for experiment 1
}
