% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_exp2.R
\name{main_function_exp2}
\alias{main_function_exp2}
\title{Function used for experiment 2}
\usage{
main_function_exp2(
  data = NULL,
  testIndexes = NULL,
  target = c("MPI", "H", "A"),
  nfolds = 5,
  methods = c("linear-pls", "beta-pls", "beta-tree-pls", "elasticnet", "beta-elastic",
    "beta-tree-elastic", "betaboost", "xgboost"),
  ...
)
}
\arguments{
\item{data}{Which of the datasets in Appendix A is preferred to used}

\item{testIndexes}{Test indexes}

\item{target}{Target variable ("MPI","H","A")}

\item{nfolds}{Number of folds used in K-Fold Cross-Validation for hyperparameter selection}

\item{methods}{Which of the methods to be used ("linear-pls","beta-pls","beta-tree-pls","elasticnet","beta-elastic","beta-tree-elastic","betaboost","xgboost")}

\item{...}{other arguments}
}
\value{
a data frame with the prediction of each method and its ground truth
}
\description{
Function used for experiment 2
}
