% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{regions_data_raw}
\alias{regions_data_raw}
\title{This is a raw dataframe that contains information about ISO codes and World Regions (it is used to add the World Regions as predictors)}
\format{
An object of class \code{data.frame} with 265 rows and 6 columns.
}
\usage{
regions_data_raw
}
\description{
This is a raw dataframe that contains information about ISO codes and World Regions (it is used to add the World Regions as predictors)
}
\author{
World Bank
}
\keyword{Bank}
\keyword{Code,}
\keyword{ISO}
\keyword{Regions,}
\keyword{World}
