% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{variables_data_raw}
\alias{variables_data_raw}
\title{This is a raw dataframe that contains variables names of the WID (it is used to remove special characters)}
\format{
An object of class \code{data.frame} with 2072 rows and 2 columns.
}
\usage{
variables_data_raw
}
\description{
This is a raw dataframe that contains variables names of the WID (it is used to remove special characters)
}
\author{
World Bank
}
\keyword{Database}
\keyword{Indicators}
\keyword{World}
