% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{webscrapped_data_experiment3}
\alias{webscrapped_data_experiment3}
\title{This is the web scrapped data for experiment 3}
\format{
An object of class \code{data.frame} with 8432 rows and 1478 columns.
}
\usage{
webscrapped_data_experiment3
}
\description{
This is the web scrapped data for experiment 3
}
\author{
Girela, Ignacio
}
\keyword{Database}
\keyword{Indicators}
\keyword{MPI,}
\keyword{Webscrapping,}
\keyword{World}
\keyword{global}
