% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{webscrapped_data_experiments1and2}
\alias{webscrapped_data_experiments1and2}
\title{This is the web scrapped data for experiments 1 and 2}
\format{
An object of class \code{data.frame} with 248 rows and 1478 columns.
}
\usage{
webscrapped_data_experiments1and2
}
\description{
This is the web scrapped data for experiments 1 and 2
}
\author{
Girela, Ignacio
}
\keyword{Database}
\keyword{Indicators}
\keyword{MPI,}
\keyword{Webscrapping,}
\keyword{World}
\keyword{global}
