% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{globalMPI_all_years_raw}
\alias{globalMPI_all_years_raw}
\title{This is the Data Table 1: National Results of the global MPI results 2022}
\format{
An object of class \code{data.frame} with 444 rows and 7 columns.
}
\usage{
globalMPI_all_years_raw
}
\description{
This is the Data Table 1: National Results of the global MPI results 2022
}
\references{
\url{https://ophi.org.uk/sites/default/files/2024-02/table-1-national-results-mpi-2022.xlsx}
}
\author{
OPHI-UNDP
}
\keyword{Multidimensional}
\keyword{Poverty}
